+++ packages/nexo_generator/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.0.7-beta.0
+
+- Version sync with nexo 0.0.7-beta.0.
+
 ## 0.0.6-beta.0
 
 - Initial release.